+++ src/config/api.js
@@ -1,10 +1,10 @@
 const baseSetup = {
-  GET_PROJECTS: 'https://bestofjs-api-v2.firebaseapp.com/',
+  GET_PROJECTS: 'https://bestofjs-api-v3.firebaseapp.com/',
   GET_README: 'https://get-github-readme-v2.now.sh',
   GET_USER_CONTENT: 'https://bestofjs-api-v1.now.sh',
   USER_CONTENT:
     'https://webtask.it.auth0.com/api/run/wt-mikeair-gmail_com-0/user-content-api-v3',
-  GET_PROJECT_DATA: 'https://bestofjs-api-v1.now.sh',
+  GET_PROJECT_DATA: 'https://bestofjs-api-v2.now.sh',
   ISSUES_REPO: 'michaelrambeau/bestofjs',
   FETCH_LICENSE: 'https://fetch-license.now.sh'
 }
@@ -12,7 +12,6 @@ const baseSetup = {
 const setup = {
   development: {
     ...baseSetup,
-    GET_PROJECTS: 'https://bestofjs-api-dev.firebaseapp.com/',
     ISSUES_REPO: 'michaelrambeau/bestofjs-sandbox'
   },
   production: {